Organize and clean up JS files
Group all 3rd-party scripts in the vendor folder and move the
other scripts to the js folder to avoid having them mixed up in
different locations (the exception is the SOAP client because it
has been modified for this project). This is also an opportunity
to remove unused dependencies (namely the combobox plugin and
the date library).

jQuery was added to the tree as a local file to avoid fetching
things from the Internet, which may not always be available (and
is not needed).

The loading of scripts now happens at the end of the body tag as
this is considered best practice. Some extra attributes that are
either unneeded or deprecated in HTML5 standard have also been
removed.