Allow single sign-on via URL params
If we want to have a link in Mantis that will redirect the user
to the board, it's useful to avoid having to log in every time.
A simple way to do that is sharing Mantis cookie with the board
by passing it via URL params, and have the board use that cookie
instead of the password (thus reusing the Mantis session).